refactor: 퀴즈 수정 관련 리팩토링, 퀴즈 정답여부 찾아오는 로직 변경

+++ src/main/java/com/twentyone/steachserver/domain/quiz/service/QuizServiceImpl.java
@@ -65,7 +65,7 @@ public class  QuizServiceImpl implements QuizService {
         }
 
         // Create and save QuizChoice entities
-        List<QuizChoice> quizChoices = quizChoiceService.createQuizChoices(choices, choices.get(answerIdx), quiz);
+        List<QuizChoice> quizChoices = quizChoiceService.createQuizChoices(choices, answerIdx, quiz);
         quiz.addChoiceList(quizChoices);
 
         return quiz;
@@ -88,17 +88,16 @@ public class  QuizServiceImpl implements QuizService {
 
     @Override
     public QuizResponseDto mapToDto(Quiz quiz) {
+        List<QuizChoice> quizChoices = quiz.getQuizChoices();
         List<String> choices = quizChoiceService.getChoices(quiz);
-        String answers = quizChoiceService.getAnswers(quiz);
 
-        int answerInt = 0;
-        for (int i =0; i<choices.size(); i++) {
-            if (choices.get(i).equals(answers)) {
-                answerInt = i+1;
+        int answerInt = 1;
+        for (QuizChoice quizChoice: quizChoices) {
+            if (quizChoice.getIsAnswer()) {
                 break;
             }
+            answerInt++;
         }
-        quizChoiceValidator.validateQuizChoices(choices, answers);
 
         return QuizResponseDto.createQuizResponseDto(quiz, choices, answerInt);
     }
@@ -124,25 +123,60 @@ public class  QuizServiceImpl implements QuizService {
 
     @Override
     @Transactional
-    public QuizResponseDto modifyQuiz(Integer quizId, QuizRequestDto dto) {
+    public QuizResponseDto modifyQuiz(Teacher teacher, Integer quizId, QuizRequestDto dto) {
+        //dto 검증로직
+        if (dto.answers() == null && dto.choices() != null) {
+            throw new IllegalArgumentException("quiz choices에 수정사항이 있다면 answer는 null이어서는 안됩니다.");
+        }
+
         Quiz quiz = quizRepository.findByIdWithQuizChoice(quizId)
                 .orElseThrow(() -> new ResourceNotFoundException("퀴즈를 찾을 수 없음"));
 
-        //검증
-        if (dto.choices() != null) {
-            List<QuizChoice> quizChoices = quiz.getQuizChoices();
-            List<String> inputChoices = dto.choices();
+        //퀴즈를 만든 사람만 수정가능
+        Teacher makeTeacher = quiz.getLecture().getCurriculum().getTeacher();
+        if (!teacher.equals(makeTeacher)) {
+            throw new ForbiddenException("퀴즈를 수정할 권한이 없습니다");
+        }
 
-            //급하니 그냥 싹 삭제.. TODO 리팩토링!!!!!!!!!
+        //answer만 바뀌는 경우
+        if (dto.choices() == null && dto.answers() != null) {
+            List<QuizChoice> quizChoices = quiz.getQuizChoices();
+            int answerIndex = dto.answers() - 1; //TODO 이건 어떻게 처리하지? 실수 발생할 것 같다..
+
+            if (answerIndex >= quiz.getQuizChoices().size() || answerIndex < 0) {
+                throw new IllegalArgumentException("정답관련 인덱스가 유효하지 않습니다.");
+            }
+
+            for (int i =0; i<quizChoices.size(); i++) {
+                if (answerIndex == i) {
+                    quizChoices.get(i).setIsAnswer();
+                } else {
+                    quizChoices.get(i).setIsNotAnswer();
+                }
+            }
+        }
+
+        //choice + answer 둘 다 바뀌는 경우
+        if (dto.choices() != null && dto.answers() != null) {
+            List<QuizChoice> quizChoices = quiz.getQuizChoices();
+
+            //TODO choice 도 몇 개 없는데, 그냥 삭제하는 게 나을까? 고민해보기..
             quiz.deleteAllQuizChoice();
             for (QuizChoice qc: quizChoices) {
                 quizChoiceService.deleteChoice(qc);
             }
 
-            quizChoiceService.createQuizChoices(inputChoices, inputChoices.get(dto.answers()-1), quiz);
-        }
+            int answerIndex = dto.answers() - 1; //TODO 이건 어떻게 처리하지? 실수 발생할 것 같다..
+            //클라이언트 인덱스 유효성 검사
+            if (answerIndex >= dto.choices().size() || answerIndex < 0) {
+                throw new IllegalArgumentException("정답관련 인덱스가 유효하지 않습니다.");
+            }
 
-        quiz.modify(dto.quizNumber(), dto.question());
+            List<String> inputChoices = dto.choices();
+            quizChoiceService.createQuizChoices(inputChoices, answerIndex, quiz);
+
+            quiz.modify(dto.quizNumber(), dto.question());
+        }
 
         return mapToDto(quiz);
     }
